requirements.txt
python-gflags<3.0
ipython
selenium
